Updated dataclass
