+++ pytools/datatools/__init__.py
@@ -1,4 +1,4 @@
-from .dataclass_response import Response, datadict
+from .datawrapper import Response, datadict
 from .dataclass_validation import validate_dataclass, validate_item
 
 dataclass = datadict
